Fix bugs and format code

+++ src/main/java/dev/hiwa/iblog/domain/dto/response/TagDto.java
@@ -18,5 +18,5 @@ public class TagDto {
     private String name;
 
     @JsonProperty(value = "post_count")
-    private Integer postCount;
+    private Long postCount;
 }
